新增供应商时添加对应价格

+++ service/src/main/java/com/huotu/hotagent/service/model/ProductPrice.java
@@ -8,10 +8,10 @@ import lombok.Data;
 @Data
 public class ProductPrice {
 
-    private int huobanMall;
-    private int dsp;
-    private int hotEdu;
-    private int thirdPartnar;
+    private double huobanMall;
+    private double dsp;
+    private double hotEdu;
+    private double thirdPartnar;
 
 
 }
